+++ src/backend/joanie/core/utils/payment_schedule.py
@@ -5,6 +5,7 @@ Payment schedule utility functions.
 import logging
 import uuid
 from datetime import date, timedelta
+from decimal import Decimal
 
 from django.conf import settings
 from django.utils import timezone
@@ -19,13 +20,14 @@ from joanie.core import enums
 from joanie.core.exceptions import InvalidConversionError
 from joanie.core.utils.emails import prepare_context_for_upcoming_installment, send
 from joanie.payment import get_country_calendar
+from joanie.payment.models import Invoice, Transaction
 
 logger = logging.getLogger(__name__)
 
 
 def _get_installments_percentages(total):
     """
-    Return the payment installments percentages for the order.
+    Return the payment installments percentages for the_ order.
     """
     percentages = None
     for limit, percentages in settings.JOANIE_PAYMENT_SCHEDULE_LIMITS.items():
@@ -238,3 +240,59 @@ def send_mail_reminder_for_installment_debit(order, installment):
             template_name="installment_reminder",
             to_user_email=order.owner.email,
         )
+
+
+def has_installment_paid(order):
+    """
+    Check if at least 1 installment is paid in the payment schedule.
+    """
+    return not order.is_free and any(
+        installment.get("state") == enums.PAYMENT_STATE_PAID
+        for installment in order.payment_schedule
+    )
+
+
+def get_paid_transactions(order):
+    """
+    Return a transactions queryset that are made from the order on paid installments.
+    """
+    return (
+        Transaction.objects.filter(
+            invoice__order=order,
+            invoice__parent__isnull=False,
+        )
+        .distinct()
+        .order_by("created_on")
+    )
+
+
+def get_transaction_references_to_refund(order) -> dict:
+    """
+    Returns a dictionary containing transaction references as key and the amount of the installment
+    that are eligible to refund in an order's payment schedule.
+    """
+    to_refund_items = {}
+    for transaction in get_paid_transactions(order):
+        for installment in order.payment_schedule:
+            if (
+                installment["state"] == enums.PAYMENT_STATE_PAID
+                and installment["amount"] == transaction.total
+            ):
+                to_refund_items[transaction.reference] = installment["amount"]
+    return to_refund_items
+
+
+def handle_refunded_transaction(invoice, amount: Decimal, refund_reference: str):
+    """
+    Handle the refund of an installment by creating a credit note, a transaction to reflect
+    the cash movement.
+    """
+    credit_note = Invoice.objects.create(
+        order=invoice.order,
+        parent=invoice.order.main_invoice,
+        total=-amount,
+        recipient_address=invoice.recipient_address,
+    )
+    Transaction.objects.create(
+        total=credit_note.total, invoice=credit_note, reference=refund_reference
+    )